CoPER_ConvE/requirements.txt
fastdtw
numpy
pyyaml
requests
six
tensorflow
tqdm
